clean environment regardless of suid status
rational: security for the auto-mounting feature, which will require
granting deck CAP_SYS_ADMIN privileges which will be inherited to mount.

To do this safely we rely on useraufs's configurable security policy,
but we need to clean the environment to make sure that we are not running
a dummy program that does not implement sanity checks.